add bitauth json files
